fix sonarqube

+++ src/main/java/br/com/ddreer/magicplaylistapi/service/ArtistServiceImpl.java
@@ -1,5 +1,6 @@
 package br.com.ddreer.magicplaylistapi.service;
 
+import br.com.ddreer.magicplaylistapi.entity.Artist;
 import br.com.ddreer.magicplaylistapi.exception.BusinessException;
 import br.com.ddreer.magicplaylistapi.model.ArtistDTO;
 import br.com.ddreer.magicplaylistapi.repository.ArtistRepository;
@@ -15,8 +16,8 @@ import java.util.UUID;
 @AllArgsConstructor
 @Log4j2
 public class ArtistServiceImpl implements ArtistService {
-    private final String NOT_FOUND = "Not found artist";
-    private final String NOT_SAVED = "Unable to save artist";
+    private static final String ARTIST_NOT_FOUND = "Not found artist";
+    private static final String ARTIST_NOT_SAVED = "Unable to save artist";
     private ArtistRepository repository;
 
     @Override
@@ -27,7 +28,7 @@ public class ArtistServiceImpl implements ArtistService {
     @Override
     public ArtistDTO findById(UUID id) {
         try {
-            return repository.findById(id).orElseThrow(() -> new BusinessException(NOT_FOUND)).toDTO();
+            return repository.findById(id).orElseThrow(() -> new BusinessException(ARTIST_NOT_FOUND)).toDTO();
         } catch (BusinessException e) {
             log.info(e.getLocalizedMessage());
             return null;
@@ -49,7 +50,7 @@ public class ArtistServiceImpl implements ArtistService {
         try {
             return repository.save(dto.toEntity()).toDTO();
         } catch (BusinessException e) {
-            log.error(NOT_SAVED);
+            log.error(ARTIST_NOT_SAVED);
             return null;
         }
     }
@@ -58,7 +59,7 @@ public class ArtistServiceImpl implements ArtistService {
     public ArtistDTO edit(ArtistDTO dto) {
         try {
             if (!repository.existsById(dto.getId())) {
-                throw new BusinessException(NOT_FOUND);
+                throw new BusinessException(ARTIST_NOT_FOUND);
             }
 
             return repository.saveAndFlush(dto.toEntity()).toDTO();
@@ -72,7 +73,9 @@ public class ArtistServiceImpl implements ArtistService {
     @Override
     public boolean delete(UUID id) {
         try {
-            repository.findById(id).orElseThrow(() -> new BusinessException(NOT_FOUND));
+            if (!repository.existsById(id)) {
+                throw new BusinessException(ARTIST_NOT_FOUND);
+            }
             repository.deleteById(id);
             return true;
         } catch (BusinessException e) {
